Fix build race conditions.
Publish will also do a build so we need to not build and publish both
otherwise we end up racing because the build and publish targets have
different global properties.

Need to fix one output directory because it builds both net461 and ns2.0
and they cannot go into the same output directory.