Update pwsh calls to follow new PS tips (#693)
In particular for cases where we are only calling a script
we want to use the Powershell filepath version instead of inline
so that we get extra logging that shows the exact command that is
being called. This helps with reproducing issues if you need to
debug.

There have been a number of attempts to add lines at the top of our
powershell script that will echo out the command line used to call it
however unfortunately all our attempts have failed for one reason or
another.

- $PSCommandPath $args -> Only shows unbound arguments
- $MyInvocation.Line -> Only shows first line so calls spread over multiple
lines fail to show the full command. See https://github.com/PowerShell/PowerShell/issues/12609

The current attempt is to simply use the DevOps task with filepath which
shows the full command line it puts in the wrapper script.