Add support for pipeline preparation in PRs. (#1104)
This pipeline adds a template to support in-situ pipeline generation from within a pull request. The model I'm going with here is that we have a common template in ```azure-sdk-tools``` under ```eng/common``` which we sync using the normal method. This template takes a repository name (defaulted), prefix and the various options for the public, internal, and test pipelines.

Then ... in each repository which supports this method of pipeline preparation (not required for iOS and C++ for example) we have another YAML file (```eng/pipelines/pipeline-preparation.yml```) which uses that template and injects any repo specific configuraiton such as the variable groups which are associated with the pipeline.

Each repo will have its own ```prepare-pipelines``` pipeline and users will be able to issue the following command to setup any missing pipelines:

```
/azp run prepare-pipelines
```

These are internal pipelines, which can be triggered only by team members with a PR comment, and they necessarily have access to the DevOps PAT. To validate that this will all hang together I did the following:

1. [Created a feature branch](https://github.com/Azure/azure-sdk-for-java/tree/feature/prepare-pipelines) on the Java repo which combines the ```eng/common``` changes from this PR, with the repo specific ```prepare-pipeline.yml``` file.
2. [Created a pipeline definition](https://dev.azure.com/azure-sdk/internal/_build?definitionId=2158&_a=summary) which points to the repo specific ```prepare-pipeline.yml``` file and set properties appropriately.
3. [Created a PR](https://github.com/Azure/azure-sdk-for-java/pull/16481) with a new service directory which tests the process end to end.

This PR lifts the changes from the feature branch that need to go into ```eng/common```. Once it is the ```eng/common``` changes from the feature branch will be removed and I'll submit another PR to add the repo-specific pipeline.

I've validated on the PR listed above that I can run the prepare pipeline successfully, and that the pipelines it generates can also be executed with an ```/azp run``` command.

That brings me to the topic of discoverability. This PR is a building block. There is another separate change that I'll be making to Check Enforcer to make this process more discoverable. The feature in Check Enforcer will be a somewhat generic timeout mechanism where when a configured timeout is reached after a PR is submitted and no check-runs have appeared, it will add a comment to the PR.

The timeout will be configurable in the CHECKENFORCER file, and the message, will be stored in a markdown configuration file alongside the CHECKENFORCER file. Check Enforcer will simply inject that markdown as a comment (once per PR, if the timeout threshold is reached). We can put whatever helpful guidance we want in that markdown. But this will be a separate PR, I want to get this basic building block into the repository first.

Here are the pipelines that were generated (will be removed once this merges):

- [java - newservice](https://dev.azure.com/azure-sdk/internal/_build?definitionId=2160)
- [java - newservice - tests](https://dev.azure.com/azure-sdk/internal/_build?definitionId=2161)
- [java - newservice - ci](https://dev.azure.com/azure-sdk/public/_build?definitionId=2159) ... failure expected.