Fix issue with .NET pipeline preparation. (#1132)
This PR fixes up a bug in pipeline preparation where we are running the tool installer in the .NET repos root directory and the dotnet tooling complaining about the invocation being ambiguous. This works around it by creating a pipeline generator path and invoking the command in that working directory rather than the source root, or the pipeline workspace.