Update Changelog Logic to accommodate older entries  (#1335)
This change is to accommodate parts of the changelog that does not properly indicate the release status as specified in the [changelog guideline](https://aka.ms/azsdk/changelogguide).
This also switched the logic to use Ordered HashTable so as not to rely on SemVer sorting. This allows us accommodate version that are not SemVer compliant.