[matrix] Fix issue where 2d index arrays with length 1 get flattened by powershell (#1521)
In order to generate sparse matrices, we generate an array of indices like `[ [0, 0], [1, 1], [2, 2] ]`. When a matrix consists of only rows/vectors with length 1, the index array will look like `[ [0, 0] ]`. Powershell will flatten the outer array of the return value to `[0, 0]` without an explicit array cast, which was missing in this case.