Set multiple sparse checkout paths in one command/job (#1577)
For pipelines with many paths (e.g. core ci), it can be annoying to have so many separate jobs show up in the UI, one for each path. Additionally, sometimes git tries to do garbage collection and it can take a lot longer to run many `git sparse-checkout add` commands as opposed to a single set command. This PR changes the behavior to parse the yaml paths parameter object into the script, and run a single command to set the sparse checkout paths. I use `invoke-expression` here because otherwise powershell treats the space-separated string as a single parameter and it all ends up on the same line in the sparse checkout config file.