Update match logic to follow standard codeowners rules
In order to help with special pipelines like aggregate-reports that live
in a general path we should follow standard codeowners rules and match
the path against the latest path in the codeowners file that matches
the file path of the pipeline. That way we can override specific pipelines
via full path later in the file to specify exact owners for that pipeline.

Add caching support for pulling definitions