Git Hub Label Management (#1829)
The focus of these changes is to update the set of tools and data
associated with managing GitHub labels across the set of Azure SDK
repositories.

The previous set of Python scripts have been replaced by Alex Ghiondea's
GHCreator tool with a set of PowerShell wrappers.  Label and repository
data are the result of recent leadership efforts to normalize labels and
reflect the currently agreed upon common set.