Move the mock attestation service to azure-sdk-tools (#1882)
## What

- Add `keyvault-mock-attestation` code to the azure-sdk-tools
- Dockerize the code to prepare for adding it to the azsdktools ACR registry.

## Why

This mock service is needed to run live tests for a specific keyvault scenario. While we currently use the 
long-running test fixture that I deployed, that is not a sustainable solution. 

We want to make sure the code for this mock service is in a repo that is owned by Azure SDK team, 
and we want to enable test runs to deploy and tear down everything that need to run live tests. Depending
on a long running test fixture that is in my private repo violates both of these conditions. By moving the 
code here and ensuring that we can push updates to ACR we will be able to deploy the container for every
test run.