Skip unnecessary and misleading role assignment warnings in local test resource deployment (#2075)
The "cannot assign owner" warning message will print on repeated local runs for the script when people don't have owner, and is somewhat of a red herring (as explained in the code comments). I don't think it's actually necessary to warn here since we already skip this section on first run of the script.