Support AAD graph and Microsoft Graph service principal APIs (#2464)
Resolves https://github.com/Azure/azure-sdk-tools/issues/2413

Until the AAD graph API is fully retired and we force local users to update to Az module 7.0.0, we need to support the breaking changes that come with Az 7.0.0 (detailed in the above issue).

Instead of trying to figure out which Az module is loaded and/or force a specific module version, this change creates a simple wrapper object for service principal creation that can handle the breaking changes between the AAD graph and MS graph objects. The two primary differences are:

1. The `ApplicationId` property becomes `AppId`
2. The service principal password is not created and returned on service principal create via the `Secret` property. Instead, you have to do a subsequent call to the `addPassword` MS Graph API. The best way I've found to do this so far (given the lack of docs) is by looking at the [service principal cmdlet](https://github.com/Azure/azure-powershell/blob/main/src/Resources/MSGraph.Autorest/custom/New-AzADSpCredential.ps1#L164). It seems like it should do this by default, but I think the parameter set requirements restrict hitting that code path?

I've tested this with Az 7.0.0 and 6.4.0 loaded, with both CI true/false and with/without cached credentials and it seems to work fine.