Update test resources SP password creation to support Az >= 7.1.0 (#2513)
Az version 7.0.0 had an issue where new service principal credential creation would not automatically generate a password by default. To work around this, one had to pass in a password credential object with an empty password to get it to work.

With the update to 7.1.0, the behavior of creating a password automatically was restored, but it also appeared to change the parameter set such that the existing calls were broken (since we weren't passing a now required `ServicePrincipalId` parameter).

```
Cannot bind argument to parameter 'ServicePrincipalId' because it is an empty string.
```

This PR updates the script to handle 7.0.0 and >= 7.1.0 versions of Az. Eventually we can remove the special handling for 7.0.0 as no one should have it installed. For posterity in case anyone hits this, the error you get when running the 7.1.0 targeted code with 7.0.0 loaded is:

```
New-TestResources.ps1: Cannot bind argument to parameter 'String' because it is null.
```

Thanks to @kashifkhan for bringing this to my attention.