Only fetch provision application oid via API if not supplied (#2592)
Currently there are some problems with the `Az.Resources` cmdlet v5.2.0 related to fetching service principals via the
graph API in private azure clouds. This manifests either as an error like below:

```
ClientSecretCredential authentication failed: AADSTS70011:
The provided request must include a 'scope' input parameter.
```

Or, if the `-MicrosoftGraphEndpointResourceId` parameter is supplied to the `Add-AzEnvironment` call, then the API
request returns a 500:

```
{"odata.error":{"code":"Request_DataContractVersionMissing","message":
{"lang":"en","value":"The specified api-version is invalid. The value must exactly match a supported version."}}}
```

I'm pretty sure there is still some missing functionality as support for non-public clouds was only recently added and
the AAD->MSGraph breaking change in azure powershell has had related issues.

Regardless of all this, we don't need to be doing an API call to get the OID in CI environments, because we already have
that value available and can pass it in as a parameter. So even if the above problems were fixed, I would still want to
make this change to reduce our test script running time. This update happens to bypass the problematic
`Get-AzAdServicePrincipal` call and fix the issue at hand.

Fixes https://github.com/Azure/azure-sdk-for-js/issues/19886