fix incidentally added path addition (#2604)
@sima-zhu I was investigating Azure/azure-sdk-for-python#16794 and discovered this.

In the case where there is a "/" in the name it gets used directly! This means on a _browser_ that the link to the landing page won't ever render because a `/` means "next folder in" on a webserver!

So for the service `Database for MySQL/PostGreSQL`, this turns into path `databaseformysql/postgresql.html`. This is 404ing.