Set owners tag in test resource deployments (#2866)
This sets up user-created resource groups to have the right metadata so that one has the option to remove DeleteAfter
tags to create long-lived resources. The alternative approach would be to update `$BaseName` generation to be formatted
like `$UserName-$ServiceDirectory` so we could parse and query the alias, but I think that would cause downstream
resource deployments that use the basename to fail validation (e.g. storage accounts don't support dash characters).