[GitHub Scripts] Milestone Dates and Repo File
The focus of these changes is to update the logic for end dates assigned
to milestones such that a milestone ends on the first Friday of the associated
month.  In addition, the set of managed repositories was added and a parameter
allowing it to be used was added to each script.