Script to verify that the json files used to generate docs have the required members. (#6045)
A while back there were changes made to the metadata json files a docs repository that caused doc entries on MsDocs to disappear and we had to ferret out where that happened. It happened because the json files contained empty versions which, ultimately, ended up causing them to get removed. This is a bare bones script that takes a few seconds to run and verifies that the json files in the metadata\latest and metadata\preview directories has the have the minimum required members and that they are non-null. It does not verify correctness of these entries. The script will scan and report all errors encountered.

Once this script is checked into common and propagated out to the language repositories, the docindex pipelines will have a task inserted after ToC generation (which can update/change the json files) and before pushing changes to the doc branch.

Ideally, a version of this script should also be run when files are directly checked into the doc repositories but the docs team does not yet have a way for us to do this and might not. The downside of this is that the docindex pipelines _could_ fail because of a previous direct check in but, even with that, it'll still identify an error which is better than hunting down why something vanished.

Fixes #5868