Replace secret rotation --expiring option with --force
The current default behavior is to rotate all secrets, including those that are not expiring, unless the --expiring option is used.
This change replaces that with a default behavior of only rotating expiring secrets unless the --force options is used, which would then also rotate the secrets that are not yet expiring.