Update all projects to v5 of graph api
Part of the tools used v5 and others v4 which causes runtime exceptions
because of the breaking changes between those version, so updated all projects
to be using v5.

Fixed bug when user doesn't have permission to update an endpoint/service connection
so it correctly errors instead of throwing null ref exception.