use azure-sdk-eng for codeowner for everything under eng folder (#8284)
Given we are moving to requiring codeowners to approve PRs I'm expanding the eng folder to the entire eng team. 

We are going to need to do that at least for eng/common in all the sync repos as well otherwise no one but you and I can merge approve a sync PR for merge.