Fix sort ordering and type alias resolution (#8438)
* Fix sort ordering and type alias resolution

Sort diagnostics and navigation slices. This requires a custom JSON
marshaler for Navigation as it contains a map and the default behavior
is non-deterministic.
Recursively resolve type aliases. This is to handle cases where a type
alias is to another type alias.

* remove custom JSON marshaler

according to the docs, the keys *are* sorted

* add unit test

* add code owners for Go APIView tool