[Validate-AzsdkCodeOwners] UX enhancements
The focus of these changes is to improve the usability and user experience of the validation script.   

The previous output was basic and lacked details outside of verbose mode - which included raw HTTP response payloads.   These changes include more context and clear color/symbol indications of what information is missing.