Use repositories file by default for GH scripts (#8527)
The default UX shouldn't change: to override any default values for disambiguating parameters, you had to pass values. If we change the default parameter set (or, for a couple scripts, leave it) to use the file we can remove the duplicate repo and language information from the `-Repositories` and `-Languages` parameters because to override those - even if they were the default parameter - you had to specify a different value. The end result should be no change.

Also cleans up warnings in the README and a couple other scripts using their respective linters I have enabled in VSCode.