Make processors transient by using a factory
Backround workers are persistent and they hold a reference to the devops processor. Processors hold VssConnections and VssConnections hold an access token.
When the token expires, the VssConnection, Processor and Background Worker need to be disposed and recreated.  This change puts a factory over the processor.
The persistent background worker creates a new process and the service provider creates a new VssConnectioin over a cached or new token.