[js-parser] Support re-exported enums (#9769)
Adds support for re-exported enums, ensuring APIView shows the correct (i.e. exported and customer-visible) 
name for an enum.

I scoped this just to enums for now, but we can expand if we find other usages.

I also went ahead and added a basic test with a basic test fixture. Future PR will work with @praveenkuttappan to ideally add this to CI builds for the parser.

Resolves #9768