Match 5 digit or more PR numbers (#10179)
The sdk automation creates branches targeting the public and private specs repos which are at PR numbers with 5 digits or longer. However during testing some get created in test-repo-billy which our clean-up doesn't know how to match but those are PR numbers are less than 5 digits so using this regex update to avoid those which will ultimately fail when we query for them.