Update Create-APIReview.ps1
Currently pipeline checks for APIView and namespace status only for all data plane releases and only .NET management plane release. We need to change this to check ONLY Namespace approval status for the management plane for Java, JS, Python and Go in addition to current logic. It should not check for APIView approval and check for namespace approval only.