Adding in some simple testing for ReadMeGenerator.
One is a unit test - pretty simple, mocks OpenAIClient and returns a faux response.

The second one is a live test, which will only run if the proper environment variables are defined. We need to get a pipeline going before that'll run anywhere else but on my local machine.

I also added in some other utilities:
- MockOutputService, if you want to just accumulate all the Output/OutputError calls into a single list and assert on it
- OpenAIMockHelper, just to make it simple to create and wire up a chat client.