Change logging functions to use Write-Host (#12532)
Change logging helpers to always write to host and either set color or use devops/gh formatting. 

We do not want to use Write-Error or Write-Warning directly because they can stop the script or not depending on preferences which makes it difficult to ensure local runs of scripts work the same as in the pipelines. So, we should never depend on these logging commands to cause a script to stop execution.